+++ src/realm/object-store/shared_realm.cpp
@@ -175,6 +175,11 @@ std::shared_ptr<AsyncOpenTask> Realm::get_synchronized_realm(Config config)
     auto coordinator = RealmCoordinator::get_coordinator(config.path);
     return coordinator->get_synchronized_realm(std::move(config));
 }
+
+std::shared_ptr<SyncSession> Realm::sync_session() const
+{
+    return m_coordinator->sync_session();
+}
 #endif
 
 void Realm::set_schema(Schema const& reference, Schema schema)